refactor(giskard-llm): streamline imports and update strict parameter in OpenAIResponseTranslator
- Replaced imports from `..utils._compact` to `..utils` in both `anthropic.py` and `azure_ai.py` for consistency.
- Modified the `strict` parameter in the tools list comprehension of `OpenAIResponseTranslator` from `True` to `None` to allow for more flexible handling of tool functions.
