+++ libs/giskard-llm/src/giskard/llm/providers/azure_ai.py
@@ -40,7 +40,7 @@ from typing import Any
 from urllib.parse import urlparse, urlunparse
 
 from ..errors import ProviderNotAvailableError
-from ..utils._compact import compact
+from ..utils import compact
 from .openai import OpenAIProvider
 
 logger = logging.getLogger(__name__)
